requirements.txt
numpy>=1.24,<3.0
PyYAML>=6.0,<7.0
matplotlib>=3.8,<4.0
pymavlink>=2.4.49,<3.0
